requirements/pc.txt
# PC control plane dependencies
jobflow>=0.1.5
jobflow-remote>=0.1.8
pymatgen>=2024.2.23
ase>=3.23.0
paramiko>=3.3.0
ruamel.yaml>=0.17.32
pandas>=2.2.0
selenium>=4.6.0
mp-api>=0.40.0
numpy>=1.26.0
